Reorganize patch files
* Move patch tests to a separate file.
* Organize types and methods.
